rename indexed to _patterns
