+++ chirho/dynamical/internals/_patterns.py
